+++ railties/lib/rails/generators/rails/app/templates/config/routes.rb
@@ -2,8 +2,9 @@
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
 
-  # You can have the root of your site routed with "root" just remember to delete public/index.html.
-  # root to: 'welcome#index'
+
+  # You can have the root of your site routed with "root"
+  root to: 'pages#index'
 
   # Example of regular route:
   #   get 'products/:id' => 'catalog#view'